Extract methods into `useInvoiceUtilities`
