+++ src/pages/invoices/routes.tsx
@@ -16,7 +16,6 @@ import { CloneInvoiceToQuote } from 'pages/quotes/clone/CloneFromInvoice';
 import { CloneInvoiceToRecurringInvoice } from 'pages/recurring-invoices/clone/CloneInvoiceToRecurringInvoice';
 import { Route } from 'react-router-dom';
 import { Clone } from './clone/Clone';
-import { Edit } from './edit/Edit';
 import { Email } from './email/Email';
 import { Invoices } from './index/Invoices';
 import { Pdf } from './pdf/Pdf';
